image_processing package renamed image_processing.py to processing.py
