+++ lcls_tools/common/image_processing/processing.py
